+++ BlazorLore.Format/BlazorLore.Format.VSCode/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to the "Blazor Formatter" extension will be documented in this file.
 
+## [1.0.1] - 2024-06-18
+
+### Added
+- Automatic CLI tool installation on first use
+- Install/Update CLI Tool command in command palette
+- Progress notifications during CLI installation
+- Automatic prompt to install CLI when formatting fails
+
+### Fixed
+- Improved handling of nested empty HTML elements
+- Fixed line break issues in element formatting
+- Better parser for complex nested structures
+
 ## [1.0.0] - 2024-06-18
 
 ### Added